+++ src/app/(route)/study/search/page.tsx
@@ -2,7 +2,8 @@ import SectionTitle from "@/common/Atoms/Text/SectionTitle";
 import StudyCardList from "@/common/Templates/CardList";
 import { getStudiesData } from "@/dummies/studies";
 import StudyCategoryTabButtonList from "../_components/StudyCategoryTabButtonList";
-import { GOALS, ONOFF } from "@/dummies/categories";
+import { GOALS } from "@/constants/categories/study_goal";
+import { ONOFF } from "@/constants/categories/study_type";
 import {
   ONOFFICONS,
   STUDYCATEGORYICONS,
@@ -57,7 +58,9 @@ export default function StudySearchPage() {
         </div>
       </div>
       <div className="flex justify-between items-end pb-6">
-        <SectionTitle size="md">전체 검색 결과 {studyCard.length}개</SectionTitle>
+        <SectionTitle size="md">
+          전체 검색 결과 {studyCard.length}개
+        </SectionTitle>
         <div className="flex gap-3 font-semibold text-sm text-[#c2c3c4]">
           <span>최신 순</span>
           <span>좋아요 순</span>
