+++ src/constants/categories/study_type.ts
@@ -1,0 +1,6 @@
+import { CategoryOption } from "@/types/model/Category";
+
+export const ONOFF: CategoryOption[] = [
+  { label: "오프라인", value: "offline" },
+  { label: "온라인", value: "online" },
+];
