+++ src/constants/menu/my_page.ts
@@ -1,0 +1,14 @@
+export const MYPAGE_MENUS = [
+  { key: "profile", href: `/my/profile`, label: "프로필" },
+  {
+    key: "study",
+    href: `/my/study`,
+    label: "내 스터디 활동",
+  },
+  {
+    key: "like-study",
+    href: `/my/like-study`,
+    label: "찜한 스터디",
+  },
+  { key: "post", href: `/my/post`, label: "내가 작성한 글" },
+];
