+++ src/dummies/categories.ts
@@ -1,224 +1,6 @@
-export const CATEGORIES = [
-  {
-    value: "cate_1",
-    label: "개발",
-    options: [
-      { value: "cate_1_1", label: "웹 개발" },
-      { value: "cate_1_2", label: "데이터 과학" },
-      { value: "cate_1_3", label: "모바일 개발" },
-      { value: "cate_1_4", label: "프로그래밍 언어" },
-      { value: "cate_1_5", label: "게임 개발" },
-      { value: "cate_1_6", label: "데이터베이스 설계 및 개발" },
-      { value: "cate_1_7", label: "스프트웨어 테스팅" },
-      { value: "cate_1_8", label: "소프트웨어 엔지니어링" },
-      { value: "cate_1_9", label: "소프트웨어 개발 도구" },
-      { value: "cate_1_10", label: "노코드 개발" },
-    ],
-  },
-  {
-    value: "cate_2",
-    label: "비즈니스",
-    options: [
-      { value: "cate_2_1", label: "기업가 정신" },
-      { value: "cate_2_2", label: "커뮤니케이션" },
-      { value: "cate_2_3", label: "경영" },
-      { value: "cate_2_4", label: "판매" },
-      { value: "cate_2_5", label: "비즈니스 전략" },
-      { value: "cate_2_6", label: "운영" },
-      { value: "cate_2_7", label: "프로젝트 관리" },
-      { value: "cate_2_8", label: "상법" },
-      { value: "cate_2_9", label: "비즈니스 분석" },
-      { value: "cate_2_10", label: "인사 관리" },
-      { value: "cate_2_11", label: "산업" },
-      { value: "cate_2_12", label: "전자 상거래" },
-      { value: "cate_2_13", label: "미디어" },
-      { value: "cate_2_14", label: "부동산" },
-      { value: "cate_2_15", label: "기타 비즈니스" },
-    ],
-  },
-  {
-    value: "cate_3",
-    label: "재무 및 회계",
-    options: [
-      { value: "cate_3_1", label: "회계 및 부기" },
-      { value: "cate_3_2", label: "준법" },
-      { value: "cate_3_3", label: "암호화폐 및 블록체인" },
-      { value: "cate_3_4", label: "경제학" },
-      { value: "cate_3_5", label: "금융" },
-      { value: "cate_3_6", label: "재무 자격증 및 시험 준비" },
-      { value: "cate_3_7", label: "금융 모델링 및 분석" },
-      { value: "cate_3_8", label: "투자 및 주식 매매" },
-      { value: "cate_3_9", label: "자금 관리 도구" },
-      { value: "cate_3_10", label: "조세" },
-      { value: "cate_3_11", label: "기타 재무 및 회계" },
-    ],
-  },
-  {
-    value: "cate_4",
-    label: "IT 및 소프트웨어",
-    options: [
-      { value: "cate_4_1", label: "IT 인증" },
-      { value: "cate_4_2", label: "네트워크 및 보안" },
-      { value: "cate_4_3", label: "하드웨어" },
-      { value: "cate_4_4", label: "운영 체제 및 서버" },
-      { value: "cate_4_5", label: "기타 IT 및 소프트웨어" },
-    ],
-  },
-  {
-    value: "cate_5",
-    label: "사무 생산성",
-    options: [
-      { value: "cate_5_1", label: "마이크로소프트" },
-      { value: "cate_5_2", label: "애플" },
-      { value: "cate_5_3", label: "구글" },
-      { value: "cate_5_4", label: "SAP" },
-      { value: "cate_5_5", label: "오라클" },
-      { value: "cate_5_6", label: "기타 사무 생산성" },
-    ],
-  },
-  {
-    value: "cate_6",
-    label: "디자인",
-    options: [
-      { value: "cate_6_1", label: "웹 디자인" },
-      { value: "cate_6_2", label: "그래픽 디자인" },
-      { value: "cate_6_3", label: "디자인 도구" },
-      { value: "cate_6_4", label: "UI/UX 디자인" },
-      { value: "cate_6_5", label: "게임 디자인" },
-      { value: "cate_6_6", label: "3D 디자인" },
-      { value: "cate_6_7", label: "애니메이션" },
-      { value: "cate_6_8", label: "패션 디자인" },
-      { value: "cate_6_9", label: "건축 디자인" },
-      { value: "cate_6_10", label: "인테리어 디자인" },
-      { value: "cate_6_11", label: "기타 디자인" },
-    ],
-  },
-  {
-    value: "cate_7",
-    label: "자기 계발",
-    options: [
-      { value: "cate_7_1", label: "자기 혁신" },
-      { value: "cate_7_2", label: "개인 생산성" },
-      { value: "cate_7_3", label: "리더십" },
-      { value: "cate_7_4", label: "경력 개발" },
-      { value: "cate_7_5", label: "육아 및 관계" },
-      { value: "cate_7_6", label: "행복" },
-      { value: "cate_7_7", label: "개인 브랜드 구축" },
-      { value: "cate_7_8", label: "창의력" },
-      { value: "cate_7_9", label: "영향력" },
-      { value: "cate_7_10", label: "자존감과 자신감" },
-      { value: "cate_7_11", label: "스트레스 관리" },
-      { value: "cate_7_12", label: "기억력 및 공부 스킬" },
-      { value: "cate_7_13", label: "동기부여" },
-      { value: "cate_7_14", label: "기타 자기 계발" },
-    ],
-  },
-  {
-    value: "cate_8",
-    label: "마케팅",
-    options: [
-      { value: "cate_8_1", label: "디지털 마케팅" },
-      { value: "cate_8_2", label: "검색 엔진 최적화" },
-      { value: "cate_8_3", label: "소셜 미디어 마케팅" },
-      { value: "cate_8_4", label: "브랜딩" },
-      { value: "cate_8_5", label: "마케팅 기초" },
-      { value: "cate_8_6", label: "마케팅 분석 및 자동화" },
-      { value: "cate_8_7", label: "홍보" },
-      { value: "cate_8_8", label: "유료 광고" },
-      { value: "cate_8_9", label: "비디오 및 모바일 마케팅 강의" },
-      { value: "cate_8_10", label: "콘텐츠 마케팅" },
-      { value: "cate_8_11", label: "그로스 해킹" },
-      { value: "cate_8_12", label: "제휴 마케팅" },
-      { value: "cate_8_13", label: "제품 마케팅" },
-      { value: "cate_8_14", label: "기타 마케팅" },
-    ],
-  },
-  {
-    value: "cate_9",
-    label: "라이프 스타일",
-    options: [
-      { value: "cate_9_1", label: "예술 및 공예" },
-      { value: "cate_9_2", label: "미용 및 화장" },
-      { value: "cate_9_3", label: "식음료" },
-      { value: "cate_9_4", label: "게임" },
-      { value: "cate_9_5", label: "주택 개조와 정원 가꾸기" },
-      { value: "cate_9_6", label: "애완동물 관리 및 훈련" },
-      { value: "cate_9_7", label: "여행" },
-      { value: "cate_9_8", label: "기타 라이프스타일" },
-    ],
-  },
-  {
-    value: "cate_10",
-    label: "사진 및 영상",
-    options: [
-      { value: "cate_10_1", label: "디지털 사진" },
-      { value: "cate_10_2", label: "사진" },
-      { value: "cate_10_3", label: "인물 사진" },
-      { value: "cate_10_4", label: "사진 도구" },
-      { value: "cate_10_5", label: "상업용 사진" },
-      { value: "cate_10_6", label: "영상 디자인" },
-      { value: "cate_10_7", label: "기타 사진 및 영상" },
-    ],
-  },
-  {
-    value: "cate_11",
-    label: "건강 및 피트니스",
-    options: [
-      { value: "cate_11_1", label: "피트니스" },
-      { value: "cate_11_2", label: "일반 건강" },
-      { value: "cate_11_3", label: "스포츠" },
-      { value: "cate_11_4", label: "영양과 식단" },
-      { value: "cate_11_5", label: "요가" },
-      { value: "cate_11_6", label: "정신 건강" },
-      { value: "cate_11_7", label: "무술과 호신술" },
-      { value: "cate_11_8", label: "안전 및 응급 처치" },
-      { value: "cate_11_9", label: "춤" },
-      { value: "cate_11_10", label: "영상" },
-      { value: "cate_11_11", label: "기타 건강 및 피트니스" },
-    ],
-  },
-  {
-    value: "cate_12",
-    label: "음악",
-    options: [
-      { value: "cate_12_1", label: "음악 프로덕션" },
-      { value: "cate_12_2", label: "음악의 기초" },
-      { value: "cate_12_3", label: "보컬" },
-      { value: "cate_12_4", label: "음악 기법" },
-      { value: "cate_12_5", label: "음악 소프트웨어" },
-      { value: "cate_12_6", label: "기타 음악" },
-    ],
-  },
-  {
-    value: "cate_13",
-    label: "교육 및 학문",
-    options: [
-      { value: "cate_13_1", label: "엔지니어링" },
-      { value: "cate_13_2", label: "인문학" },
-      { value: "cate_13_3", label: "수학" },
-      { value: "cate_13_4", label: "과학" },
-      { value: "cate_13_5", label: "온라인 교육" },
-      { value: "cate_13_6", label: "사회 과학" },
-      { value: "cate_13_7", label: "언어 배우기" },
-      { value: "cate_13_8", label: "강사 연수" },
-      { value: "cate_13_9", label: "시험 준비" },
-      { value: "cate_13_10", label: "기타 교육 및 학문" },
-    ],
-  },
-];
-
-export const GOALS = [
-  { label: "개념학습", value: "goal_1" },
-  { label: "응용/활용", value: "goal_2" },
-  { label: "프로젝트", value: "goal_3" },
-  { label: "자격증/시험", value: "goal_4" },
-  { label: "취업/면접", value: "goal_5" },
-  { label: "챌린지", value: "goal_6" },
-  { label: "특강", value: "goal_7" },
-  { label: "취미", value: "goal_8" },
-];
-
-export const ONOFF = [
-  { label: "오프라인", value: "offline" },
-  { label: "온라인", value: "online" },
-];
+/**
+ * *Move to dir @constants/categories*
+ * CATEGORIES --> /classified_job.ts
+ * GOALS --> /study_goal.ts
+ * ONOFF --> /study_type.ts
+ */
