+++ src/types/model/Category.ts
@@ -1,0 +1,10 @@
+export type CategoryGroup = {
+  value: string;
+  label: string;
+  options: CategoryOption[];
+};
+
+export type CategoryOption = {
+  value: string;
+  label: string;
+};
